+++ src/de/tilman_neumann/jml/factor/lehman/Lehman_Simple.java
@@ -52,10 +52,10 @@ public class Lehman_Simple extends FactorAlgorithmBase {
 		for (int k=1; k <= cbrt; k++) {
 			long fourKN = k*N<<2;
 			double fourSqrtK = Math.sqrt(k<<4);
-			int sqrt4kN = (int) Math.ceil(Math.sqrt(fourKN)); // ceil() is required for stability
-			int limit = (int) (sqrt4kN + sixthRoot / fourSqrtK);
-			for (int a = sqrt4kN; a <= limit; a++) {
-				long test = a*(long)a - fourKN;
+			long sqrt4kN = (long) Math.ceil(Math.sqrt(fourKN)); // ceil() is required for stability
+			long limit = (long) (sqrt4kN + sixthRoot / fourSqrtK);
+			for (long a = sqrt4kN; a <= limit; a++) {
+				long test = a*a - fourKN;
 				long b = (long) Math.sqrt(test);
 				if (b*b == test) {
 					return gcdEngine.gcd(a+b, N);
